修复学分修改问题

+++ src/main/java/com/Main/RowMapper/CourseRowMapper.java
@@ -14,7 +14,7 @@ public class CourseRowMapper implements RowMapper<Course> {
         course.setName(rs.getString("course_name"));
         course.setDescription(rs.getString("course_description"));
         course.setCategory(rs.getString("category"));
-        course.setCredit(rs.getInt("credit"));
+        course.setCredit(rs.getFloat("credit"));
         course.setTeacherId(rs.getInt("teacher_id"));
         course.setHours_per_week(rs.getInt("hours_per_week"));
         return course;
